+++ radio/src/telemetry/telemetry.cpp
@@ -20,6 +20,7 @@
 
 #include "opentx.h"
 #include "multi.h"
+#include "mixer_scheduler.h"
 
 uint8_t telemetryStreaming = 0;
 uint8_t telemetryRxBuffer[TELEMETRY_RX_PACKET_SIZE];   // Receive buffer. 9 bytes (full packet), worst case 18 bytes with byte-stuffing (+1)
